// FE/src/api/models/practitionerDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * SIGREF API
 * API de Auditoría para Sistema de Gestión de Referencias FHIR
 * OpenAPI spec version: 1.0.0
 */
import type { IdentifierDto } from "./identifierDto";
import type { HumanNameDto } from "./humanNameDto";
import type { ContactPointDto } from "./contactPointDto";
import type { AdministrativeGender } from "./administrativeGender";
import type { PractitionerRoleDto } from "./practitionerRoleDto";

export interface PractitionerDto {
  /** @nullable */
  id?: string | null;
  /** @nullable */
  identifier?: IdentifierDto[] | null;
  /** @nullable */
  active?: boolean | null;
  /** @nullable */
  name?: HumanNameDto[] | null;
  /** @nullable */
  telecom?: ContactPointDto[] | null;
  gender?: AdministrativeGender;
  /** @nullable */
  birthDate?: string | null;
  /** @nullable */
  lastUpdated?: string | null;
  /** @nullable */
  roles?: PractitionerRoleDto[] | null;
}