// FE/src/api/models/practitionerRoleDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SIGREF.API
 * OpenAPI spec version: 1.0
 */
import type { IdentifierDto } from "./identifierDto";
import type { PeriodDto } from "./periodDto";
import type { ReferenceDto } from "./referenceDto";
import type { CodeableConceptDto } from "./codeableConceptDto";
import type { ExtendedContactDetailDto } from "./extendedContactDetailDto";
import type { AvailabilityDto } from "./availabilityDto";

export interface PractitionerRoleDto {
  /** @nullable */
  id?: string | null;
  /** @nullable */
  identifier?: IdentifierDto[] | null;
  /** @nullable */
  active?: boolean | null;
  period?: PeriodDto;
  practitioner?: ReferenceDto;
  organization?: ReferenceDto;
  /** @nullable */
  network?: ReferenceDto[] | null;
  /** @nullable */
  code?: CodeableConceptDto[] | null;
  /** @nullable */
  specialty?: CodeableConceptDto[] | null;
  /** @nullable */
  location?: ReferenceDto[] | null;
  /** @nullable */
  healthcareService?: ReferenceDto[] | null;
  /** @nullable */
  contact?: ExtendedContactDetailDto[] | null;
  /** @nullable */
  characteristic?: CodeableConceptDto[] | null;
  /** @nullable */
  communication?: CodeableConceptDto[] | null;
  availability?: AvailabilityDto;
  /** @nullable */
  endpoint?: ReferenceDto[] | null;
  /** @nullable */
  display?: string | null;
  /** @nullable */
  lastUpdated?: string | null;
}